Finished batch for pump and start for sonar
Finally did some work on the bash script for the pumping of water. This
goes along a systemd unit which I'll incorporate later to the project.

Started on good pace with the distance measuring probe. Now I've just
got to iron the wrinkles on the initial noise, and add the adjustment
(if there's any).